+++ src/main/java/gift/web/dto/request/product/CreateProductRequest.java
@@ -1,5 +1,6 @@
-package gift.web.dto.request;
+package gift.web.dto.request.product;
 
+import gift.domain.Product;
 import gift.web.validation.constraints.RequiredKakaoApproval;
 import gift.web.validation.constraints.SpecialCharacter;
 import jakarta.validation.constraints.NotBlank;
@@ -7,7 +8,7 @@ import java.net.URL;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.Range;
 
-public class UpdateProductRequest {
+public class CreateProductRequest {
 
     @NotBlank
     @Length(min = 1, max = 15)
@@ -18,7 +19,7 @@ public class UpdateProductRequest {
     private final Integer price;
     private final URL imageUrl;
 
-    public UpdateProductRequest(String name, Integer price, URL imageUrl) {
+    public CreateProductRequest(String name, Integer price, URL imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -34,5 +35,13 @@ public class UpdateProductRequest {
 
     public URL getImageUrl() {
         return imageUrl;
+    }
+
+    public Product toEntity() {
+        return new Product.Builder()
+            .name(this.name)
+            .price(this.price)
+            .imageUrl(this.imageUrl)
+            .build();
     }
 }
