+++ src/main/java/gift/web/dto/request/product/UpdateProductRequest.java
@@ -1,6 +1,5 @@
-package gift.web.dto.request;
+package gift.web.dto.request.product;
 
-import gift.domain.Product;
 import gift.web.validation.constraints.RequiredKakaoApproval;
 import gift.web.validation.constraints.SpecialCharacter;
 import jakarta.validation.constraints.NotBlank;
@@ -8,7 +7,7 @@ import java.net.URL;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.Range;
 
-public class CreateProductRequest {
+public class UpdateProductRequest {
 
     @NotBlank
     @Length(min = 1, max = 15)
@@ -19,7 +18,7 @@ public class CreateProductRequest {
     private final Integer price;
     private final URL imageUrl;
 
-    public CreateProductRequest(String name, Integer price, URL imageUrl) {
+    public UpdateProductRequest(String name, Integer price, URL imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -35,13 +34,5 @@ public class CreateProductRequest {
 
     public URL getImageUrl() {
         return imageUrl;
-    }
-
-    public Product toEntity() {
-        return new Product.Builder()
-            .name(this.name)
-            .price(this.price)
-            .imageUrl(this.imageUrl)
-            .build();
     }
 }
