+++ src/main/java/gift/web/dto/response/product/ReadProductResponse.java
@@ -1,25 +1,24 @@
-package gift.web.dto.response;
+package gift.web.dto.response.product;
 
 import gift.domain.Product;
 import java.net.URL;
 
-public class UpdateProductResponse {
+public class ReadProductResponse {
 
     private final Long id;
     private final String name;
     private final Integer price;
     private final URL imageUrl;
 
-    private UpdateProductResponse(Long id, String name, Integer price, URL imageUrl) {
+    private ReadProductResponse(Long id, String name, Integer price, URL imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
     }
 
-    public static UpdateProductResponse from(Product product) {
-        return new UpdateProductResponse(product.getId(), product.getName(), product.getPrice(),
-            product.getImageUrl());
+    public static ReadProductResponse fromEntity(Product product) {
+        return new ReadProductResponse(product.getId(), product.getName(), product.getPrice(), product.getImageUrl());
     }
 
     public Long getId() {
@@ -37,4 +36,5 @@ public class UpdateProductResponse {
     public URL getImageUrl() {
         return imageUrl;
     }
+
 }
